OPHSUPA-63 Mallinnettu vapaan sivistystyön suoritukset UI:lle

+++ suorituspalvelu-shared/src/main/scala/fi/oph/suorituspalvelu/parsing/koski/KoskiParser.scala
@@ -37,7 +37,7 @@ case class Arviointi(arvosana: KoskiKoodi,
                      `päivä`: Option[String],
                      hyväksytty: Boolean)
 
-case class Laajuus(arvo: Int,
+case class Laajuus(arvo: BigDecimal,
                    yksikkö: Option[Yksikko])
 
 case class OsaSuoritus(tyyppi: SuoritusTyyppi,
